+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed:
+- Navbar changed to Salomon style
+
+### Fixed:
+- Navigation glitch where the bottom nav bar was moving
+- Error handling when page does not exist
+
+## [0.8.126] - 2022-08-08
+### Changed:
 - Updated dependencies
 - Inline code style in editor